[refactor] 패키지 이동 - Pawn, PawnColor 를 pieces 패키지로 이동
