+++ chess/src/main/java/chess/Board.java
@@ -1,5 +1,7 @@
 package chess;
 
+import pieces.Pawn;
+
 import java.util.ArrayList;
 import java.util.List;
