+++ chess/src/test/java/pieces/PawnTest.java
@@ -1,7 +1,8 @@
-package chess;
+package pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
